admin style

+++ Server/model/traineesModel.js
@@ -1,4 +1,6 @@
 const pool = require('../DB.js');
+const nodemailer = require('nodemailer');
+require('dotenv').config();
 
 async function getAllTrainees() {
     try {
@@ -82,7 +84,7 @@ async function deleteWaitingTrainees(query) {
     }
 }
 
-async function deleteTrainee(id) {
+async function deleteTrainee(id, sendMail) {
     const connection = await pool.getConnection();
     try {
         await connection.beginTransaction();
@@ -93,6 +95,13 @@ async function deleteTrainee(id) {
             throw new Error('Trainee not found');
         }
         const info_id = infoIdResult[0].information_id;
+
+        const userSql = `select email, first_name from users where user_id = ?`;
+        const [userResult] = await connection.query(userSql, [id]);
+        if (userResult.length === 0) {
+            throw new Error('User not found');
+        }
+        const user = userResult[0];
 
         const traineeInClassSql = `delete from trainees_in_class where trainee_id = ?`;
         await connection.query(traineeInClassSql, [id]);
@@ -109,10 +118,13 @@ async function deleteTrainee(id) {
         const traineeSql = `delete from trainees where trainee_id = ?`;
         await connection.query(traineeSql, [id]);
 
-        const userSql = `delete from users where user_id = ?`;
-        await connection.query(userSql, [id]);
+        const deleteUserSql = `delete from users where user_id = ?`;
+        await connection.query(deleteUserSql, [id]);
 
         await connection.commit();
+
+        if (sendMail) sendEmailToUser(user);
+
         return { success: true, message: "Trainee deleted successfully" };
     } catch (err) {
         await connection.rollback();
@@ -122,6 +134,44 @@ async function deleteTrainee(id) {
         connection.release();
     }
 }
+
+
+
+const { SENDER_EMAIL, APP_PASSWORD } = process.env
+
+const sendMail = async (transporter, mailOptions) => {
+
+    try {
+        await transporter.sendMail(mailOptions);
+        console.log('Email sent');
+    }
+    catch (error) {
+        console.error(error);
+    }
+}
+
+const transporter = nodemailer.createTransport({
+    service: "gmail",
+    host: "smtp.gmail.com",
+    port: 587,
+    secure: false,
+    auth: {
+        user: SENDER_EMAIL,
+        pass: APP_PASSWORD,
+    },
+});
+
+const sendEmailToUser = (user) => {
+    const mailOptions = {
+        from: SENDER_EMAIL,
+        to: user.email,
+        subject: `Hi ${user.first_name}`,
+        text: `Due to unusual behavior that does not comply with the regulations of the Fithub website, we are forced to delete your account. Your account has been deleted. FitHub`
+    }
+    sendMail(transporter, mailOptions)
+}
+
+
 
 
 async function updateTrainee(body, id) {
